+++ src/main/java/com/olivera/sistema_reparacion/application/usecases/reparacion/EliminarReparacionPorIdImpl.java
@@ -2,6 +2,7 @@ package com.olivera.sistema_reparacion.application.usecases.reparacion;
 
 import com.olivera.sistema_reparacion.application.ports.in.reparacion.EliminarReparacionPorId;
 import com.olivera.sistema_reparacion.application.ports.out.ReparacionRepositoryPort;
+import com.olivera.sistema_reparacion.domain.exceptions.reparacion.ReparacionNoEncontradaException;
 
 public class EliminarReparacionPorIdImpl implements EliminarReparacionPorId {
 
@@ -13,6 +14,9 @@ public class EliminarReparacionPorIdImpl implements EliminarReparacionPorId {
 
     @Override
     public void eliminarReparacionPorId(Long id) {
+        if(!reparacionRepositoryPort.existsById(id)){
+            throw new ReparacionNoEncontradaException("Reparacion no encontrada");
+        }
         reparacionRepositoryPort.eliminarReparacionPorId(id);
     }
 }
